refactor: move the rule timing table formatter to internal/output
